requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-appconf==1.0.6
django-debug-toolbar==4.4.6
django-filter==24.2
django-imagekit==5.0.0
django-jazzmin==3.0.0
django-modeltranslation==0.19.4
django-solo==2.3.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
flower==2.0.1
humanize==4.10.0
inflection==0.5.1
kombu==5.3.7
openpyxl==3.1.5
packaging==24.1
pilkit==3.0
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
